requirements.txt
torch>=2.0.0
transformers>=4.28.0
diffusers>=0.15.0
accelerate>=0.19.0
pillow
numpy
scipy
flux-1-dev
streamlit
